Move src/elona/version.* to src/version.*
With this change, the object file generated from version.cpp is
independent of libelona. Before, libelona was linked from scratch since
version.cpp was automatically generated and updated every build by
CMake. This change will speed up each build time.
