+++ src/version.cpp.in
